requirements.txt
pure-python-adb==0.3.0.dev0
ollama==0.1.8
python-dotenv==0.20.0
